requirements.txt
flask
asyncio
uvicorn
openai
python-dotenv
semantic-kernel
azure-search-documents
fastapi
requests
matplotlib
chainlit